Removed unnecessary casting

+++ src/ConverterTest.java
@@ -26,14 +26,14 @@ public class ConverterTest {
 
     @Test
     public void testToInteger() throws Exception {
-        assertEquals(0, (int) converter.stringToInt("0"));
-        assertEquals(1, (int) converter.stringToInt("1"));
-        assertEquals(-1, (int) converter.stringToInt("-1"));
-        assertEquals(1001, (int) converter.stringToInt("1001"));
-        assertEquals(987654321, (int) converter.stringToInt("987654321"));
-        assertEquals(-123456789, (int) converter.stringToInt("-123456789"));
-        assertEquals(001, (int) converter.stringToInt("001"));
-        assertEquals(10001, (int) converter.stringToInt("10001"));
+        assertEquals(0, converter.stringToInt("0"));
+        assertEquals(1, converter.stringToInt("1"));
+        assertEquals(-1, converter.stringToInt("-1"));
+        assertEquals(1001, converter.stringToInt("1001"));
+        assertEquals(987654321, converter.stringToInt("987654321"));
+        assertEquals(-123456789, converter.stringToInt("-123456789"));
+        assertEquals(001, converter.stringToInt("001"));
+        assertEquals(10001, converter.stringToInt("10001"));
         assertEquals(null, converter.stringToInt(null));
     }
